+++ src/assets/scss/default.scss
@@ -95,7 +95,6 @@ body {
 	display: flex;
 	flex-direction: column;
 	flex-grow: 1;
-	margin-top: 4.5rem;
 	.bouncy {
 		display: flex;
 		flex-direction: column;
