+++ src/components/base/Chatsen.vue
@@ -1,5 +1,5 @@
 <template>
-	<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 288 288" width="64" height="64">
+	<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 288 288" width="1em">
 		<path
 			fill="currentColor"
 			d="M145.18,288c-69.19-.16-125.36-46.79-140.91-109.18C.89,165.23-.56,151.42,.2,137.44c1.99-36.78,15.96-68.39,41.77-94.64C61.22,23.22,84.36,10.29,111.09,3.86c8.2-1.97,16.51-3.1,24.93-3.61,14.69-.88,29.14,.54,43.38,4.16,6.08,1.55,11.42,4.67,16.31,8.51,11.95,9.4,18.88,21.7,20.64,36.79,.41,3.55,.25,7.18,.11,10.77-.15,3.73-3.4,5.71-6.97,4.42-2.48-.9-4.92-1.95-7.43-2.75-7.93-2.54-16.04-4.1-24.41-3.81-17.72,.62-32.67,7.49-44.57,20.57-9.11,10.01-15.94,21.43-19.37,34.67-7.09,27.43,2.55,56.19,24.65,73.96,16.63,13.37,27.43,30.44,31.88,51.35,3.26,15.32,2.4,30.48-2.31,45.43-.39,1.22-.93,1.91-2.32,2.1-7.14,.97-14.3,1.51-20.44,1.59Z"
